fix: separated helper functions
